Generate proper meta-data
To avoid that the app.* keys are shown as unknown in the editor, we can
enable the annotation processor.

Spring Boot metadata annotation processor looks for
`@ConfigurationProperties` bean on class and `@Bean` definition. For the
latter it exposes the properties found on the return type of the method.

Previously, we exposed `javax.sql.DataSource` that does not expose any
property (getter/setter) so the meta-data for that is empty. We switched
to the Tomcat connection pool to expose its relevant properties.